src/requirements.txt
django==4.0.5
gunicorn
django-cors-headers
djangorestframework==3.13.1
psycopg2-binary
pytest-django==4.5.2
channels
daphne
channels_redis
PyJWT
mixer
pydantic==1.10.5
factory_boy
python-memcached
